requirements.txt
# Core dependencies
fastapi
uvicorn
python-multipart    
websockets

# LLM and AI
google-generativeai

# Voice Processing
speechrecognition
pydub
pyaudio
elevenlabs
openai-whisper

# Google Calendar Integration
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# Utilities
python-dotenv
httpx
asyncio-mqtt
dateparser
pytz
tzlocal
sqlalchemy
aiosqlite

# Audio processing
numpy
scipy
librosa

# Development
pytest
pytest-asyncio
black
flake8 